Remove example page
