+++ client/src/routes/index.svelte
@@ -1,17 +1,80 @@
 <script>
-  import ExampleComponent from "../components/ExampleComponent.svelte";
+  let src =
+    "https://www.emoji.com/wp-content/uploads/filebase/thumbnails/icons/emoji-icon-glossy-00-05-faces-face-fantasy-smiling-poo-with-heart-eyes-72dpi-forPersonalUseOnly.png";
+
+  function handleRegister() {
+    alert("regiter clicked");
+  }
+  function handleLogin() {
+    alert("login");
+  }
 </script>
 
-<div class:centerer={true}>
-  <!-- Here's how you could extract conditional class groups with @apply -->
-  <ExampleComponent
-    title="🌐 Sapper with PostCSS project base"
-    paragraph="This is an example route and component to make sure everything's working."
-  />
-</div>
+<div
+  class="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8"
+>
+  <div class="max-w-md w-full space-y-8">
+    <div>
+      <img class="mx-auto h-35 w-auto" {src} alt="Workflow" />
+      <h2 class="mt-6 text-center text-3xl font-extrabold text-gray-900">
+        Sign in or register
+      </h2>
+    </div>
+    <form class="mt-8 space-y-6" action="#" method="POST">
+      <input type="hidden" name="remember" value="true" />
+      <div class="rounded-md shadow-sm -space-y-px">
+        <div>
+          <label for="email-address" class="sr-only">Email address</label>
+          <input
+            id="email-address"
+            name="email"
+            type="email"
+            autocomplete="email"
+            required
+            class="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded-t-md focus:outline-none focus:ring-yellow-500 focus:border-yellow-700 focus:z-10 sm:text-sm"
+            placeholder="Email address"
+          />
+        </div>
+        <div>
+          <label for="password" class="sr-only">Password</label>
+          <input
+            id="password"
+            name="password"
+            type="password"
+            autocomplete="current-password"
+            required
+            class="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded-b-md focus:outline-none focus:ring-yellow-500 focus:border-yellow-700 focus:z-10 sm:text-sm"
+            placeholder="Password"
+          />
+        </div>
+      </div>
 
-<style>
-  .centerer {
-    @apply flex-1 flex flex-col items-center justify-center;
-  }
-</style>
+      <div class="flex items-center justify-between">
+        <div class="flex items-center">
+          <input
+            id="remember_me"
+            name="remember_me"
+            type="checkbox"
+            class="h-4 w-4 text-indigo-600 focus:ring-yellow-600 border-gray-300 rounded"
+          />
+          <label for="remember_me" class="ml-2 block text-sm text-gray-900">
+            Remember me
+          </label>
+        </div>
+      </div>
+
+      <div>
+        <button
+          type="submit"
+          class="my-1 group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-yellow-800 hover:bg-yellow-900 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-yellow-400">
+          Sign in
+        </button>
+        <button
+          type="submit"
+          class="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-yellow-800 hover:bg-yellow-900 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-yellow-400">
+          Register
+        </button>
+      </div>
+    </form>
+  </div>
+</div>
